+++ ScoutSuite/frontend/src/partials/aws/s3.buckets.id/Informations/index.js
